requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pywin32>=305
psutil>=5.9.0
